+++ utils.h
@@ -8,16 +8,16 @@
 typedef Adafruit_SH1106G OLED;
 
 struct Vector2 {
-  double x;
-  double y;
+  float x;
+  float y;
 
-  Vector2(double x = 0, double y = 0)
+  Vector2(float x = 0, float y = 0)
     : x(x), y(y) {}
-  
-  void set(double x, double y);
+
+  void set(float x, float y);
   void set(const Vector2& other);
-  void from_angle(double angle);
-  double magnitude();
+  void from_angle(float angle);
+  float magnitude();
 };
 
 struct Size {
@@ -26,11 +26,11 @@ struct Size {
 
   Size(uint16_t width = 0, uint16_t height = 0)
     : width(width), height(height) {}
-  
+
   void set(uint16_t width, uint16_t height);
 };
 
 Size get_text_size(OLED& oled, const String& text);
-double clamp(double value, double minimum, double maximum);
+float clamp(float value, float minimum, float maximum);
 
 #endif
